feat: 상위 카테고리 조회API 구현

+++ RisingCamp-Spring-Template/src/main/java/com/example/demo/src/domain/entitiy/Address.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class Address {
-    private int addressIdx;
+    private Long addressIdx;
     private String name;
     private String phoneNumber;
     private String city;
